Convert all Integers to Long for future support of 53/54 bit Whole Numbers defined by stix 2.1 spec

+++ src/main/java/io/digitalstate/stix/coo/extension/types/HttpRequestExtensionExt.java
@@ -60,7 +60,7 @@ public interface HttpRequestExtensionExt extends CyberObservableExtension {
     //@TODO Review if this should be a long
     @JsonProperty("message_body_length")
     @JsonPropertyDescription("Specifies the length of the HTTP message body, if included, in bytes.")
-    Optional<Integer> getMessageBodyLength();
+    Optional<Long> getMessageBodyLength();
 
     /*
      * Must be of type artifact
